+++ R/bbk-package.R
@@ -22,7 +22,7 @@
 #' @importFrom httr2 resp_status
 #' @importFrom httr2 resp_url
 #' @importFrom httr2 resps_data
-#' @importFrom stats na.omit setNames
+#' @importFrom stats median na.omit setNames
 "_PACKAGE"
 
 col_order = c("date", "key", "value", "freq", "title", "description")